Make library descriptions paragraphs
This is one of the _very few_ things about markdown that actually
bothers me. By the standard spec, when you would like to insert a line
break in a paragraph you must end the line with two trailing spaces.
This is in general _OK_ (Though arguments can certainly be made against
it), but the real problem is when using git it will ignore this trailing
white space (mainly when adding via `git add -p`)...

This may actually be semantically better...